requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
boto3==1.28.21
botocore==1.31.21
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.3
Django==4.1.7
django-cors-headers==3.14.0
django-debug-toolbar==4.1.0
django-environ==0.10.0
django-extensions==3.2.3
django-redis==5.3.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.0
drf-yasg==1.21.5
et-xmlfile==1.1.0
frozenlist==1.4.0
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.3
MarkupSafe==2.1.2
multidict==6.0.4
numpy==1.25.1
openai==0.27.8
openpyxl==3.1.2
packaging==23.0
pandas==2.0.3
Pillow==10.0.0
pycparser==2.21
PyJWT==2.6.0
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
redis==5.0.0
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.3
swagger-spec-validator==3.0.3
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.15
yarl==1.9.2
